Support gitlab on checkout_pr
